Use task return values for success / fail
Allow the task to indicate success or failure with a return value.

This means the function doesn't have to result in an abnormal shutdown
for the task to be retried.

Functions that return `:ok` or `{:ok, result}` are considered
successful. Any other return value is considered a failure and will
cause the task to be retried.

ENSI-167 #comment Task can indicate error with return value